// src/generated/t_tech/invest/grpc/marketdata.ts
export enum TradeSourceType {
  /** TRADE_SOURCE_UNSPECIFIED - Тип источника сделки не определён. */
  TRADE_SOURCE_UNSPECIFIED = 0,
  /** TRADE_SOURCE_EXCHANGE - Биржевые сделки. */
  TRADE_SOURCE_EXCHANGE = 1,
  /** TRADE_SOURCE_DEALER - Сделки дилера. */
  TRADE_SOURCE_DEALER = 2,
  /** TRADE_SOURCE_ALL - Все сделки. */
  TRADE_SOURCE_ALL = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum TradeDirection {
  /** TRADE_DIRECTION_UNSPECIFIED - Направление сделки не определено. */
  TRADE_DIRECTION_UNSPECIFIED = 0,
  /** TRADE_DIRECTION_BUY - Покупка. */
  TRADE_DIRECTION_BUY = 1,
  /** TRADE_DIRECTION_SELL - Продажа. */
  TRADE_DIRECTION_SELL = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface Trade {
  /** FIGI-идентификатор инструмента. */
  figi: string;
  /** Направление сделки. */
  direction: TradeDirection;
  /** Цена за 1 инструмент. Чтобы получить стоимость лота, нужно умножить на лотность инструмента. [Подробнее про перевод цен в валюту](./faq_marketdata/#_15). */
  price:
    | Quotation
    | undefined;
  /** Количество лотов. */
  quantity: number;
  /** Время сделки в часовом поясе UTC по времени биржи. */
  time:
    | Date
    | undefined;
  /** UID инструмента. */
  instrumentUid: string;
  /** Тип источника сделки. */
  tradeSource: TradeSourceType;
  /** Тикер инструмента. */
  ticker: string;
  /** Класс-код (секция торгов). */
  classCode: string;
}